+++ xarray/core/indexing.py
@@ -512,6 +512,7 @@ class DaskIndexingAdapter(utils.NDArrayMixin):
     def __getitem__(self, key):
         def to_int_tuple(key):
             # workaround for uint64 indexer (GH:1406)
+            # TODO remove here after next dask releas (0.15.3)
             return tuple([k.astype(int) if isinstance(k, np.ndarray)
                           else k for k in key])