+++ doc/whats-new.rst
@@ -120,6 +120,10 @@ Bug fixes
   objects with data stored as ``dask`` arrays (:issue:`1529`).
   By `Joe Hamman <https://github.com/jhamman>`_.
 
+- :py:func:`~xarray.open_rasterio` method now skips rasterio.crs -attribute if
+  it is none.
+  By `Leevi Annala <https://github.com/leevei>`_.
+
 .. _whats-new.0.9.6:
 
 v0.9.6 (8 June 2017)