+++ doc/whats-new.rst
@@ -16,7 +16,7 @@ What's New
 .. _whats-new.0.9.7:
 
 v0.10.0 (unreleased)
--------------------
+--------------------
 
 Backward Incompatible Changes
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -27,7 +27,7 @@ Backward Incompatible Changes
   indexing, as well as vectorized indexing.
   Due to this change, existing uses of xarray objects to index other xarray
   objects will break in some cases.
-  See :ref:`indexing` for the details
+  See :ref:`vectorized_indexing` for the details
   (:issue:`1444`, :issue:`1436`, ).
   By `Keisuke Fujii <https://github.com/fujiisoup>`_ and
   `Stephan Hoyer <https://github.com/shoyer>`_.